Add implementation to find key into aws with parameter key_name
