--- result.txt
@@ -1,20 +1,0 @@
-
-================================================================================
-Wazuh VM Test Summary
-================================================================================
-
-Overall Status: PASS
-
-Total Tests: 1
-Passed: 1
-Failed: 0
-Warnings: 0
-Errors: 0
-Skipped: 0
-
---------------------------------------------------------------------------------
-Passed Tests
---------------------------------------------------------------------------------
-
-Certificates Module:
-  ✓ Certificates: Certificates exist
