--- result3.md
@@ -1,47 +1,0 @@
-# Wazuh VM Test Results
-
-## Summary
-
-**Status**: PASS :green_circle:
-
-| Metric | Count |
-|--------|-------|
-| Total Tests | 16 |
-| Passed | 16 |
-| Failed | 0|
-| Warnings | 0 |
-| Skipped | 0 |
-
-## Passed Tests 
-
-### Certificates
-
-- Certificates: Certificates exist :green_circle:
-- Certificates: Certificates validity :green_circle:
-- Certificates: Certificate subjects :green_circle:
-- Certificates: Certificate issuers :green_circle:
-
-### Connectivity
-
-- Connectivity: Service connectivity :green_circle:
-- Connectivity: Wazuh api connectivity :green_circle:
-
-### Logs
-
-- Logs: Log files exist :green_circle:
-- Logs: Logs for errors :green_circle:
-- Logs: Recent logs :green_circle:
-
-### Services
-
-- Services: Services active :green_circle:
-- Services: Services running :green_circle:
-- Services: Required directories :green_circle:
-- Services: Required files :green_circle:
-- Services: Ports listening :green_circle:
-- Services: Health endpoints :green_circle:
-
-### Version
-
-- Version: Services versions :green_circle:
-
