+++ wazuh_vm_tester/src/vm_tester/__init__.py
@@ -4,6 +4,7 @@ Wazuh VM Tester - Framework for validating Wazuh VMs using pytest.
 
 __version__ = "0.2.0"
 
-from .config import AMITesterConfig, setup_logging, get_logger
+from .config import AMITesterConfig
+from .utils.logger import setup_logging, get_logger
 
 __all__ = ["AMITesterConfig", "setup_logging", "get_logger"]
