+++ wazuh_vm_tester/src/vm_tester/conftest.py
@@ -12,7 +12,8 @@ src_dir = Path(__file__).parent.parent
 if str(src_dir) not in sys.path:
     sys.path.insert(0, str(src_dir))
 
-from .config import AMITesterConfig, get_logger
+from .config import AMITesterConfig
+from .utils.logger import get_logger
 from .connections.pytest_connector import get_connection
 
 logger = get_logger(__name__)
