+++ wazuh_vm_tester/src/vm_tester/instances/factory.py
@@ -4,7 +4,8 @@ Factory module for creating instances.
 
 from typing import Optional
 
-from ..config import AMITesterConfig, get_logger
+from ..config import AMITesterConfig
+from ..utils.logger import get_logger
 from .base import InstanceInterface
 from .local_instance import LocalInstance
 from .ec2_instance import EC2Instance
