+++ wazuh_vm_tester/src/vm_tester/reporting/base.py
@@ -6,7 +6,7 @@ from enum import Enum
 from typing import Dict, List, Any
 from datetime import datetime
 
-from ..config import get_logger
+from ..utils.logger import get_logger
 
 logger = get_logger(__name__)
