+++ wazuh_vm_tester/src/vm_tester/tests/test_certificates.py
@@ -6,7 +6,8 @@ import os
 import pytest
 from datetime import datetime
 
-from ..config import AMITesterConfig, get_logger
+from ..config import AMITesterConfig
+from ..utils.logger import get_logger
 from ..connections.pytest_connector import get_connection
 
 logger = get_logger(__name__)
