+++ wazuh_vm_tester/src/vm_tester/tests/test_version.py
@@ -7,7 +7,8 @@ import pytest
 import re
 from typing import Dict, List, Optional
 
-from ..config import AMITesterConfig, get_logger
+from ..config import AMITesterConfig
+from ..utils.logger import get_logger
 from ..connections.pytest_connector import get_connection
 
 logger = get_logger(__name__)
